+++ test/test_models.py
@@ -1,39 +1,11 @@
-import unittest
 import json
 import peewee
 from unittest.mock import patch
 
-from khux_medal_finder.models import BaseModel, Medal
 from khux_medal_finder.factories import MedalFactory
+from khux_medal_finder.models import BaseModel, Medal
 
-test_db = peewee.SqliteDatabase(':memory:')
-MODELS = [BaseModel, Medal]
-
-class BaseDBTestCase(unittest.TestCase):
-    """TestCase class to use in the test cases where we need to query a DB.
-    It uses a temporary DB to avoid messing with our production one"""
-
-    def setUp(self):
-
-        # Bind model classes to test db
-        for model in MODELS:
-            model.bind(test_db, bind_refs=False, bind_backrefs=False)
-
-        test_db.connect()
-        test_db.create_tables(MODELS)
-
-        super(BaseDBTestCase, self).setUp()
-
-    def tearDown(self):
-        # Not strictly necessary since in-memory databases only live
-        # for the duration of the connection, and in the next step we close
-        # the connection...but a good practice all the same.
-        test_db.drop_tables(MODELS)
-
-        # Close connection to db.
-        test_db.close()
-
-        super(BaseDBTestCase, self).tearDown()
+from test.helpers import BaseDBTestCase
 
 class TestModels(BaseDBTestCase):
     """Tests for the different ORM models"""
